Reorganize and add alternative to handling propTypes
